* 'release' of https://github.com/0AndWild/Hanghae99-Dogfaw:
  postService 게시글수정(수정) - 업데이트시 stack부분 업데이트 반영이 되지않아 수정함
  키워드 - SSE 기능 작업중 (mypageService, userApplicationService 아직 적용X........)
  마이페이지 작성한글 불러오기(추가) - 리스폰스값에 해당 게시글 현재인원,최대인원, 지원자 수 추가
  AcceptanceService 신청수락 (추가) - 신청 수락 로직에서 신청 수락 시 현재인원 +1 된 후 현재인원과 모집마감인원이 동일할 경우 deadLine = true로 변경되게함
  1. UserApplicationService 지원하기 신청/취소(추가) - 현재 모집글 신청 수락된 유저일 경우 다시 지원을 하지 못하게 로직 추가 2. Post - 작성자기준 모집마감,모집마감취소 Api 추가
  마이페이지 지원자보기, 내팀원보기(수정) - reponse값 stack부분이 모두 동일하게가는 부분 수정하여 이제 각 유저의 스택이 정상적으로 잘 보내짐
  마이페이지(기능)- 내팀원보기, 팀원추방하기, 팀 탈퇴하기 추가
  키워드 - 충돌처리 및 SSE패키지 생성
  Post Delete- 수락기능이 추가되어 게시글 삭제시 게시글과 연관된 수락 테이블도 삭제
  키워드 -  댓글조회 예외처리
  mypageController 유저 프로필 이미지 업로드 -  기존 이미지 유지 또는 기본이미지(null) 일경우 백에서는 MultipartFile(required=false) 로 한 후 프론트에서 요청 시 body에 해당하는 부분만 데이터를 요청하면 multipart를 null로 인식해줌(문제해결!)
  mypageController 유저 프로필 이미지 업로드 - multipartFile null 값을 못받아 String image = multipartFile.getContentType(); 을 이용해 null일 경우는 contnetType도 null로 받아 시도
  mypage controller 프로필편집 -  닉네임, 기술스택만 편집할 때  if (multipartFile!=null) 로 테스트
  mypage 프로필이미지 수정 -MypageController 프로필이미지 업로드  if(multipartFile.isEmpty()) -> if(multipartFile.getSize()==0)으로 변경

+++ src/main/java/com/project/dogfaw/acceptance/AcceptanceService.java
@@ -42,6 +42,10 @@ public class AcceptanceService {
         //신청자확인
         userApplicationRepository.findByUserAndPost(applier,post)
                 .orElseThrow(()-> new CustomException(ErrorCode.APPLY_NOT_FOUND));
+        //수락완료 저장 전 현재모집인원 체크
+        if(post.getCurrentMember()>=post.getMaxCapacity()){
+            throw new CustomException(ErrorCode.ACCEPTANCE_PEOPLE_SET_CLOSED);
+        }
         //신청상태 삭제
         userApplicationRepository.deleteByUserAndPost(applier,post)
                 .orElseThrow(()-> new CustomException(ErrorCode.APPLY_NOT_FOUND));
@@ -49,17 +53,24 @@ public class AcceptanceService {
         acceptanceRepository.save(acceptance);
         //현재인원+1
         post.increaseCnt(); //post.decreaseCnt = 수락된 신청자가 상세페이지에서 참여취소 하였을 때 acceptanceRepo에서 삭제하고 -1 해야함
+        //모집정원이 모두 찼을 경우 모집마감
+        if(post.getCurrentMember()==post.getMaxCapacity()){
+            Boolean deadline = true;
+            post.updateDeadline(deadline);
+        }
 
     }
     @Transactional
     /*신청거절하기*/
     public void rejection(Long userId, Long postId, User user) {
+        //신청자
         User applier = userRepository.findById(userId)
                 .orElseThrow(()->new CustomException(ErrorCode.SIGNUP_USERID_NOT_FOUND));
+        //게시글
         Post post = postRepository.findById(postId)
                 .orElseThrow(()->new CustomException(ErrorCode.POST_NOT_FOUND));
+        //작성자
         Long writer = post.getUser().getId();
-
 
         //작성자확인
         if(!writer.equals(user.getId())){
